requirements.txt
google-generativeai>=0.8.0
Pillow>=10.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
